refactor(store,runtime,tools): drop redundancy the audit flagged
Five independent leftovers, none behaviour-changing except the first.

`snapshot-payload-builder` read every `turn_results` row for a session to
build a plugin-id set, then filtered `plugin_data` by it. The filter could
never drop anything: the set is the union of both sources, so every
plugin_data row's id is in it by construction — the comment above it said as
much. Long sessions paid a full-table read (each row carrying its complete
runtimeResults payload) for a no-op. Now it just reads the plugin data.

`summarizeStorageMigrations(registry = REGISTRY)` returned its argument, and
no caller ever passed one. It is a constant, so it is now `STORAGE_MIGRATIONS`.

`vector-factory.ts` held nine lines of type aliases and a comment explaining
that there is no vector-store factory. Merged into `vector-store.ts`.

`TimeCursor` was defined twice — identically — in `@covel/store` and
`@covel/shared`. Store already depends on shared, so it re-exports that one.

`ResolvedTool` / `ToolClient` / `ToolDefinition` had no consumer in or out of
the tools package (`ResolvedTool.requiresApproval` was never read; the
`.requiresApproval` reads in the tree are all `PluginTrustInfo`'s). `client.ts`
keeps `ToolCallResult`, which the runtime does use.
