+++ packages/store/src/vector-store.ts
@@ -176,3 +176,11 @@ export function supportsVector(
     typeof candidate.resolveSessionVectorTarget === "function"
   );
 }
+
+/**
+ * Vector backend selector and the composed vector-store shape.
+ * Production discovers vector support via `supportsVector(store)` directly
+ * (see vector-store.ts) — there is no runtime vector-store factory.
+ */
+export type VectorBackend = "embedded" | "none" | "external";
+export type VectorStore = VectorStoreCapability & VectorModelOps;
