+++ packages/tools/src/client.ts
@@ -1,22 +1,9 @@
 /**
- * Unified tool client contract.
+ * Result shape a tool execution can return: the bare output, or an envelope
+ * carrying pending proposals / emitted events alongside it.
  */
 
-import type { ToolExecutionContext, ResolvedTool } from "./types.js";
 import type { ToolExecutionEnvelope } from "./result.js";
-
-export type ToolDefinition = ResolvedTool;
 
 export type ToolCallResult<TOutput = unknown> =
   TOutput | ToolExecutionEnvelope<TOutput>;
-
-export interface ToolClient {
-  readonly id: string;
-  list(): Promise<readonly ToolDefinition[]>;
-  call(
-    name: string,
-    args: unknown,
-    ctx: ToolExecutionContext,
-  ): Promise<ToolCallResult>;
-  close?(): Promise<void>;
-}
